fix: rename SidebarSearchFriend -> SidebarSearchUsers
