+++ src/components/modules/SidebarDashboard/SidebarDashboard.tsx
@@ -6,7 +6,7 @@ import SidebarContacts from '../SidebarContacts/SidebarContacts';
 import SidebarGroups from '../SidebarGroups/SidebarGroups';
 import SidebarChats from '../SidebarChats/SidebarChats';
 import SidebarSettings from '../SidebarSettings/SidebarSettings';
-import SidebarSearchFriend from '../SidebarSearchFriend/SidebarSearchFriend';
+import SidebarSearchUsers from '../SidebarSearchUsers/SidebarSearchUsers';
 
 import './SidebarDashboard.scss';
 
@@ -19,7 +19,7 @@ const SidebarDashboard: React.FC = () => {
       {navbarAction === 1 && <SidebarChats />}
       {navbarAction === 2 && <SidebarGroups />}
       {navbarAction === 3 && <SidebarContacts />}
-      {navbarAction === 4 && <SidebarSearchFriend />}
+      {navbarAction === 4 && <SidebarSearchUsers />}
       {navbarAction === 5 && <SidebarSettings />}
     </div>
   );
